UGS,EUS,stock and more

+++ app/src/main/java/com/example/a963103033239757ba10504dc3857ddc7/data/model/StationModel.kt
@@ -11,8 +11,8 @@ data class StationModel(
     @PrimaryKey val name: String,
     var coordinateX: Double,
     var coordinateY: Double,
-    var capacity: Long,
-    var stock: Long,
-    var need: Long,
+    var capacity: Int,
+    var stock: Int,
+    var need: Int,
     var isFav: Boolean = false
 )
